thay đổi loại biến int của GIANHAP và GIABAN thành float

+++ QuanLyPMT/BUS/Thuoc_BUS.cs
@@ -14,11 +14,11 @@ namespace BUS
         {
             return thuoc_DAL.getDataThuoc();
         }
-        public void addThuoc(string TENTHUOC, string DONVT, string DUONGDUNG, int TONKHO, int GIANHAP, int GIABAN, string NHACUNGCAP, string GHICHU)
+        public void addThuoc(string TENTHUOC, string DONVT, string DUONGDUNG, int TONKHO, float GIANHAP, float GIABAN, string NHACUNGCAP, string GHICHU)
         {
             thuoc_DAL.addThuoc(TENTHUOC, DONVT, DUONGDUNG, TONKHO, GIANHAP, GIABAN, NHACUNGCAP, GHICHU);
         }
-        public void updateThuoc(int MATHUOC, string TENTHUOC, string DONVT, string DUONGDUNG, int TONKHO, int GIANHAP, int GIABAN, string NHACUNGCAP, string GHICHU)
+        public void updateThuoc(int MATHUOC, string TENTHUOC, string DONVT, string DUONGDUNG, int TONKHO, float GIANHAP, float GIABAN, string NHACUNGCAP, string GHICHU)
         {
             thuoc_DAL.updateThuoc(MATHUOC, TENTHUOC, DONVT, DUONGDUNG, TONKHO, GIANHAP, GIABAN, NHACUNGCAP, GHICHU);
         }
