+++ QuanLyPMT/DTO/Thuoc.cs
@@ -11,8 +11,8 @@ namespace DTO
         private string donVT;
         private string duongDUNG;
         private int tonKHO;
-        private int giaNHAP;
-        private int giaBAN;
+        private float giaNHAP;
+        private float giaBAN;
         private string nhaCUNGCAP;
         private string ghiCHU;
         public int MATHUOC { get => maTHUOC; set => maTHUOC = value; }
@@ -20,8 +20,8 @@ namespace DTO
         public string DONVT { get => donVT; set => donVT = value; }
         public string DUONGDUNG { get => duongDUNG; set => duongDUNG = value; }
         public int TONKHO { get => tonKHO; set => tonKHO = value; }
-        public int GIANHAP { get => giaNHAP; set => giaNHAP = value; }
-        public int GIABAN { get => giaBAN; set => giaBAN = value; }
+        public float GIANHAP { get => giaNHAP; set => giaNHAP = value; }
+        public float GIABAN { get => giaBAN; set => giaBAN = value; }
         public string NHACUNGCAP { get => nhaCUNGCAP; set => nhaCUNGCAP = value; }
         public string GHICHU { get => ghiCHU; set => ghiCHU = value; }
     }
